api/pyproject.toml
[project]
name = "api"
version = "1.0.0"
description = "Django REST API for automated security scanning of GitHub repositories to detect exposed secrets and sensitive information"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "celery>=5.5.3",
    "cryptography>=46.0.3",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.1",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
    "requests>=2.32.5",
    "watchdog>=6.0.0",
]
